requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
pillow-heif>=0.8.0
pytesseract>=0.3.8
requests>=2.25.0
onnx>=1.10.0
onnxruntime>=1.8.0
numpy>=1.21.0
